+++ boris-core/boris-core.cabal
@@ -40,6 +40,7 @@ library
                     Boris.Core.Data.Instance
                     Boris.Core.Data.Log
                     Boris.Core.Data.Time
+                    Boris.Core.Data.Workspace
 
                     Boris.Core.Serial.Ref
                     Boris.Core.Serial.Command
